同意各种方式请求下单数据 完善下单请求类的构造函数(加入无参构造函数)与注释(主要是提示可为null)
